+++ dev/com.ibm.ws.microprofile.reactive.streams.operators/test/src/com/ibm/ws/microprofile/reactive/streams/test/ErrorRecoveryTest.java
@@ -12,8 +12,6 @@
 
 package com.ibm.ws.microprofile.reactive.streams.test;
 
-import java.util.concurrent.ExecutionException;
-
 import org.eclipse.microprofile.reactive.streams.operators.ProcessorBuilder;
 import org.eclipse.microprofile.reactive.streams.operators.PublisherBuilder;
 import org.eclipse.microprofile.reactive.streams.operators.ReactiveStreams;
@@ -21,10 +19,18 @@ import org.junit.Test;
 
 import io.reactivex.subscribers.TestSubscriber;
 
+/**
+ * A class that unit tests errors, recovery handling and fallback
+ */
 public class ErrorRecoveryTest extends WASReactiveUT {
 
+    /**
+     * A user's Processor throws an exception
+     *
+     * @throws InterruptedException
+     */
     @Test
-    public void testError() throws InterruptedException, ExecutionException {
+    public void testError() throws InterruptedException {
 
         PublisherBuilder<Integer> data = ReactiveStreams.of(1, 2, 3, 4, 5);
         ProcessorBuilder<Integer, Integer> errorInjector = ReactiveStreams.<Integer>builder().
@@ -45,8 +51,13 @@ public class ErrorRecoveryTest extends WASReactiveUT {
                 .assertValueAt(1, 2);
     }
 
+    /**
+     * Tests catching exceptions with onErrorResume operator
+     *
+     * @throws InterruptedException
+     */
     @Test
-    public void testErrorRecovery() throws InterruptedException, ExecutionException {
+    public void testErrorRecovery() throws InterruptedException {
 
         PublisherBuilder<String> data = ReactiveStreams.of("tick", "tick", "boom", "tick");
         ProcessorBuilder<String, String> errorInjector = ReactiveStreams.<String>builder().
@@ -73,8 +84,14 @@ public class ErrorRecoveryTest extends WASReactiveUT {
 
     }
 
+    /**
+     * Tests catching exceptions with onErrorResume operator plus splicing on a
+     * continuation of the stream with concat
+     *
+     * @throws InterruptedException
+     */
     @Test
-    public void testConcatFallback() throws InterruptedException, ExecutionException {
+    public void testConcatFallback() throws InterruptedException {
 
         PublisherBuilder<String> data1 = ReactiveStreams.of("tick", "tick", "boom", "tick");
         PublisherBuilder<String> fallback = ReactiveStreams.of("TICK");
